personalized_grpo_tutor/requirements.txt
torch>=2.3.0
transformers>=4.46.0
datasets>=2.21.0
accelerate>=0.34.0
peft>=0.13.0
trl>=0.16.1
bitsandbytes>=0.43.3
pyyaml>=6.0.2
tqdm>=4.66.5
numpy>=1.26.4
